pyproject.toml
[tool.poetry]
name = "invest-bot"
version = "0.1.0"
description = "invest bot for auto trading with Tinkoff Investment"
authors = ["smeyanoff <miradg99@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "invest_bot"}]

[tool.poetry.dependencies]
python = "^3.8"
tinkoff-investments = "^0.2.0b46"
pandas = "^1.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
